{"schema":"decodex/commit/1","summary":"Split markdown header summary modules","authority":"manual"}
